refactor(clawdstrike-policy-event): extract kinds/policy_event from edr/receipt/mod.rs
