+++ crates/libs/clawdstrike-policy-event/src/edr/receipt/kinds/mod.rs
@@ -3,6 +3,7 @@ mod evidence_bundle;
 mod graph_slice;
 mod policy_decision;
 mod policy_delta;
+mod policy_event;
 mod sensor_state;
 mod simulation;
 
@@ -11,5 +12,6 @@ pub(crate) use evidence_bundle::*;
 pub(crate) use graph_slice::*;
 pub(crate) use policy_decision::*;
 pub(crate) use policy_delta::*;
+pub(crate) use policy_event::*;
 pub(crate) use sensor_state::*;
 pub(crate) use simulation::*;
